// old/src/app/models/external/client/cliente-farmacorp-dto.ts
export interface ClienteFarmacorpDto {
    idShopify?: string;
    idCliente?: string;
    idGrupo?: string;
    idCodigo?: string;
    nroEmpleado?: string;
    nombres?: string;
    apellidos?: string;
    cedula?: string;
    keySexo?: string;
    sexo?: string;
    fechaNacimiento?: string;
    email?: string;
    celular?: string;
    telefono?: string;
    contacto?: string;
    orgName?: string;
    businessName?: string;
    orgTypeCode?: string;
    orgTypeDesc?: string;
    keyTipo?: string;
    estado?: number;
}